requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
geopandas==0.14.1
shapely==2.0.2
rasterio==1.3.9
openai==1.3.7
chromadb==0.4.18
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0